pyproject.toml
[tool.poetry]
name = "image-man"
version = "0.1.0"
description = ""
authors = ["Joseph Banks <joseph@josephbanks.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.1"
pyyaml = "^5.1"
attrdict = "^2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
